+++ NOTICE
@@ -1,4 +1,4 @@
-Splunk SOAR SlackBot
+Splunk SOAR Slackbot
 Copyright (c) 2022 Splunk Inc.
 
 Third-party Software Attributions:
